third_party/WebKit/LayoutTests/fast/css/invalidation/explicit-inheritance-flag-expected.txt

Explicit inheritance of non-inherited property. Test that the explicitInherit flag in ComputedStyle is maintained correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(root.querySelector('span')).maxHeight is "100px"
PASS successfullyParsed is true

TEST COMPLETE
